requirements.txt
tkinterdnd2
pillow
moviepy
pillow-avif-plugin
pyinstaller